Version 2.6
Added compatibility to ULX, This Scoreboard now works with Exsto and ULX
(maybe evolve in the future).
Organized files.
Added some Lua coding conventions.
Added a little of Commenting to the Source Code.
Ajusted Labels and columns.
It is compatible with ULX and uTime or only ULX.
Added Creative Commons Attribution-NonCommercial-ShareAlike 4.0
International License.
Lots of improvement and fixes.